+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/DependencyManager.groovy
@@ -8,6 +8,9 @@ import life.qbic.business.project.load.LoadProjectsOutput
 import life.qbic.business.samples.count.CountSamples
 import life.qbic.business.samples.count.CountSamplesDataSource
 import life.qbic.business.samples.count.CountSamplesOutput
+import life.qbic.business.samples.download.DownloadSamples
+import life.qbic.business.samples.download.DownloadSamplesDataSource
+import life.qbic.business.samples.download.DownloadSamplesOutput
 import life.qbic.datamodel.dtos.portal.PortalUser
 import life.qbic.datamodel.dtos.projectmanagement.Project
 import life.qbic.portal.sampletracking.communication.notification.MessageBroker
@@ -17,6 +20,8 @@ import life.qbic.portal.sampletracking.components.projectoverview.CountSamplesPr
 import life.qbic.portal.sampletracking.components.projectoverview.LoadProjectsPresenter
 import life.qbic.portal.sampletracking.components.projectoverview.ProjectOverviewView
 import life.qbic.portal.sampletracking.components.projectoverview.ProjectOverviewViewModel
+import life.qbic.portal.sampletracking.components.projectoverview.download.DownloadProjectController
+import life.qbic.portal.sampletracking.components.projectoverview.download.ManifestPresenter
 import life.qbic.portal.sampletracking.datasources.Credentials
 import life.qbic.portal.sampletracking.datasources.OpenBisConnector
 import life.qbic.portal.sampletracking.datasources.database.DatabaseSession
@@ -46,6 +51,7 @@ class DependencyManager {
 
     private LoadProjectsDataSource loadProjectsDataSource
     private CountSamplesDataSource countSamplesDataSource
+    private DownloadSamplesDataSource downloadSamplesDataSource
 
     private ResourceService<Project> projectResourceService
     private ResourceService<StatusCount> statusCountService
@@ -86,6 +92,7 @@ class DependencyManager {
         DatabaseSession.init(user, password, host, port, sqlDatabase)
         SamplesDbConnector samplesDbConnector = new SamplesDbConnector(DatabaseSession.getInstance())
         countSamplesDataSource = samplesDbConnector
+        downloadSamplesDataSource = samplesDbConnector
 
         Credentials openBisCredentials = new Credentials(
                 user: configurationManager.getDataSourceUser(),
@@ -118,8 +125,17 @@ class DependencyManager {
      */
     private ProjectOverviewView createProjectOverviewView() {
         ProjectOverviewViewModel viewModel = new ProjectOverviewViewModel(projectResourceService, statusCountService)
-        ProjectOverviewView view =  new ProjectOverviewView(viewModel)
+        
+        DownloadProjectController downloadController = setupDownloadProjectUsecase(viewModel)
+        ProjectOverviewView view =  new ProjectOverviewView(notificationService, viewModel, downloadController)
         return view
+    }
+    
+    private DownloadProjectController setupDownloadProjectUsecase(ProjectOverviewViewModel viewModel) {
+        DownloadSamplesOutput output = new ManifestPresenter(notificationService, viewModel)
+        DownloadSamples downloadSamples = new DownloadSamples(downloadSamplesDataSource, output)
+        
+        return new DownloadProjectController(downloadSamples)
     }
 
     /**
@@ -144,7 +160,7 @@ class DependencyManager {
         projectCodes.each {
             countSamples.countReceivedSamples(it)
             countSamples.countQcFailedSamples(it)
-            countSamples.countAvailableDataSamples(it)        
+            countSamples.countAvailableDataSamples(it)
         }
     }