+++ sample-tracking-status-overview-domain/src/main/groovy/life/qbic/business/samples/count/CountSamples.groovy
@@ -6,7 +6,7 @@ import life.qbic.datamodel.samples.Status
 /**
  * <b>Count samples</b>
  *
- * <p>This use case counts samples of a project and the subset of samples that have been received at the lab.</p>
+ * <p>This use case counts samples of a project and the subset of samples that have a specific status.</p>
  *
  * @since 1.0.0
  */
@@ -40,22 +40,29 @@ class CountSamples implements CountSamplesInput{
   void countReceivedSamples(String projectCode) {
     try {
       sampleStatuses = dataSource.fetchSampleStatusesForProject(projectCode)
+      // counts samples that have AT LEAST this status (or a later one)
       int receivedAmount = countSamplesFromStatus(Status.SAMPLE_RECEIVED)
       output.countedReceivedSamples(projectCode, sampleStatuses.size(), receivedAmount)
     } catch (DataSourceException dataSourceException) {
       output.failedExecution(dataSourceException.getMessage())
-    } catch (Exception e) {
+    } catch (Exception ignored) {
       output.failedExecution("Could not count received samples.")
     }
   }
 
+  /**
+   * This method calls the output interface with the number of all samples in a project as
+   * well as the subset of samples which had a failed quality control
+   * In case of failure the output interface failure method is called.
+   * @since 1.0.0
+   */
   @Override
   void countQcFailedSamples(String projectCode) {
     try {
       sampleStatuses = dataSource.fetchSampleStatusesForProject(projectCode)
       int receivedAmount = sampleStatuses.findAll { it == Status.SAMPLE_QC_FAIL }.size()
       output.countedFailedQcSamples(projectCode, sampleStatuses.size(), receivedAmount)
-    }catch (Exception e) {
+    }catch (Exception ignored) {
       output.failedExecution("Could not count failed qc samples.")
     }
   }